Seperated helper functions and interfaces for simplicity in backend
