+++ apps/server/src/modules/technicians/index.ts
@@ -6,6 +6,7 @@ export {
   type UpdateTechnicianData,
   type TechnicianProfile,
   type TechnicianProfileRow,
+  type TechnicianListDTO,
 } from './technicians.repository.js';
 export { TechniciansService, type ITechniciansService } from './technicians.service.js';
 export { TechniciansController } from './technicians.controller.js';
