+++ apps/server/src/modules/technicians/technicians.controller.ts
@@ -1,12 +1,13 @@
 import type { Request, Response } from 'express';
 import type { ITechniciansService } from './technicians.service.js';
+import { parseCoords } from '../../shared/utils/technicians/index.js';
 
 export class TechniciansController {
   constructor(private readonly service: ITechniciansService) {}
 
   async getByCategoryId(req: Request, res: Response): Promise<void> {
     const categoryId = req.params.categoryId as string;
-    const { lat, lng } = this.parseCoords(req);
+    const { lat, lng } = parseCoords(req);
     const technicians = await this.service.getTechniciansByCategory(categoryId, lat, lng);
     res.json({ technicians });
   }
@@ -18,21 +19,9 @@ export class TechniciansController {
       res.status(400).json({ error: 'Query parameter "q" is required' });
       return;
     }
-    const { lat, lng } = this.parseCoords(req);
+    const { lat, lng } = parseCoords(req);
     const technicians = await this.service.searchTechniciansByCategory(categoryId, query, lat, lng);
     res.json({ technicians });
-  }
-
-  private parseCoords(req: Request): { lat?: number; lng?: number } {
-    const rawLat = req.query.lat as string | undefined;
-    const rawLng = req.query.lng as string | undefined;
-    if (rawLat == null || rawLng == null) return {};
-
-    const lat = Number.parseFloat(rawLat);
-    const lng = Number.parseFloat(rawLng);
-    if (Number.isNaN(lat) || Number.isNaN(lng)) return {};
-
-    return { lat, lng };
   }
 
   async getProfile(req: Request, res: Response): Promise<void> {
@@ -46,4 +35,3 @@ export class TechniciansController {
     }
   }
 }
-
