+++ apps/server/src/shared/utils/technicians/index.ts
@@ -1,2 +1,3 @@
+export { parseCoords } from './parseCoords.js';
 export { distanceKm } from './geo.js';
 export { sortByDistance } from './sorting.js';
